requirements.txt
Flask
requests
openai
beautifulsoup4
pytesseract
Pillow
python-dotenv
requests-html
selenium
apscheduler
backoff
